Refactor `zcashd_regtest_controller::with_temporary_node`; improve error message clarity.
